+++ cache/in-memory/src/event/guild.rs
@@ -18,6 +18,7 @@ impl InMemoryCache {
         if self.wants(ResourceType::CHANNEL) {
             self.0.guild_channels.insert(guild.id, HashSet::new());
             self.cache_guild_channels(guild.id, guild.channels);
+            self.cache_guild_channels(guild.id, guild.threads);
         }
 
         if self.wants(ResourceType::EMOJI) {
@@ -216,7 +217,10 @@ impl UpdateCache for GuildUpdate {
 mod tests {
     use super::*;
     use twilight_model::{
-        channel::{ChannelType, GuildChannel, TextChannel},
+        channel::{
+            thread::{AutoArchiveDuration, PublicThread, ThreadMember, ThreadMetadata},
+            ChannelType, GuildChannel, TextChannel,
+        },
         guild::{
             DefaultMessageNotificationLevel, ExplicitContentFilter, MfaLevel, NSFWLevel,
             PartialGuild, Permissions, PremiumTier, SystemChannelFlags, VerificationLevel,
@@ -239,6 +243,35 @@ mod tests {
             position: 1,
             rate_limit_per_user: None,
             topic: None,
+        })]);
+
+        let threads = Vec::from([GuildChannel::PublicThread(PublicThread {
+            id: ChannelId(222),
+            default_auto_archive_duration: None,
+            guild_id: None,
+            kind: ChannelType::GuildPublicThread,
+            last_message_id: None,
+            message_count: 0,
+            name: "guild thread with no guild id".to_owned(),
+            owner_id: None,
+            parent_id: None,
+            rate_limit_per_user: None,
+            member_count: 0,
+            thread_metadata: ThreadMetadata {
+                archived: false,
+                auto_archive_duration: AutoArchiveDuration::Hour,
+                archive_timestamp: "".to_string(),
+                invitable: None,
+                locked: false,
+            },
+            member: Some(ThreadMember {
+                flags: 0,
+                id: Some(ChannelId(1)),
+                join_timestamp: "".to_string(),
+                member: None,
+                presence: None,
+                user_id: Some(UserId(2)),
+            }),
         })]);
 
         let guild = Guild {
@@ -278,6 +311,7 @@ mod tests {
             system_channel_id: None,
             system_channel_flags: SystemChannelFlags::SUPPRESS_JOIN_NOTIFICATIONS,
             rules_channel_id: None,
+            threads,
             unavailable: false,
             verification_level: VerificationLevel::VeryHigh,
             voice_states: Vec::new(),
@@ -293,6 +327,7 @@ mod tests {
         cache.cache_guild(guild);
 
         let channel = cache.guild_channel(ChannelId(111)).unwrap();
+        let thread = cache.guild_channel(ChannelId(222)).unwrap();
 
         // The channel was given to the cache without a guild ID, but because
         // it's part of a guild create, the cache can automatically attach the
@@ -300,6 +335,13 @@ mod tests {
         // correct value.
         match channel {
             GuildChannel::Text(ref c) => {
+                assert_eq!(Some(GuildId(123)), c.guild_id);
+            }
+            _ => panic!("{:?}", channel),
+        }
+
+        match thread {
+            GuildChannel::PublicThread(ref c) => {
                 assert_eq!(Some(GuildId(123)), c.guild_id);
             }
             _ => panic!("{:?}", channel),
@@ -349,6 +391,7 @@ mod tests {
             stickers: Vec::new(),
             system_channel_flags: SystemChannelFlags::empty(),
             system_channel_id: None,
+            threads: Vec::new(),
             unavailable: false,
             vanity_url_code: None,
             verification_level: VerificationLevel::VeryHigh,