+++ cache/in-memory/src/event/mod.rs
@@ -10,6 +10,7 @@ pub mod reaction;
 pub mod role;
 pub mod stage_instance;
 pub mod sticker;
+pub mod thread;
 pub mod voice_state;
 
 use crate::{config::ResourceType, InMemoryCache, UpdateCache};