+++ http/src/routing/route_display.rs
@@ -69,6 +69,20 @@ impl Display for RouteDisplay<'_> {
 
                 Display::fmt(role_id, f)
             }
+            Route::AddThreadMember {
+                channel_id,
+                user_id,
+            }
+            | Route::RemoveThreadMember {
+                channel_id,
+                user_id,
+            } => {
+                f.write_str("channels/")?;
+                Display::fmt(channel_id, f)?;
+                f.write_str("/thread-members/")?;
+
+                Display::fmt(user_id, f)
+            }
             Route::CreateBan {
                 guild_id,
                 delete_message_days,
@@ -245,6 +259,23 @@ impl Display for RouteDisplay<'_> {
                 Display::fmt(guild_id, f)?;
 
                 f.write_str("/templates")
+            }
+            Route::CreateThread { channel_id } => {
+                f.write_str("channels/")?;
+                Display::fmt(channel_id, f)?;
+
+                f.write_str("/threads")
+            }
+            Route::CreateThreadFromMessage {
+                channel_id,
+                message_id,
+            } => {
+                f.write_str("channels/")?;
+                Display::fmt(channel_id, f)?;
+                f.write_str("/messages/")?;
+                Display::fmt(message_id, f)?;
+
+                f.write_str("/threads")
             }
             Route::CreateTypingTrigger { channel_id } => {
                 f.write_str("channels/")?;
@@ -491,6 +522,11 @@ impl Display for RouteDisplay<'_> {
                 token,
                 webhook_id,
             }
+            | Route::GetFollowupMessage {
+                application_id: webhook_id,
+                interaction_token: token,
+                message_id,
+            }
             | Route::GetWebhookMessage {
                 message_id,
                 token,
@@ -544,6 +580,12 @@ impl Display for RouteDisplay<'_> {
                 Display::fmt(channel_id, f)?;
 
                 f.write_str("/followers")
+            }
+            Route::GetActiveThreads { guild_id } => {
+                f.write_str("guild_id/")?;
+                Display::fmt(guild_id, f)?;
+
+                f.write_str("/threads/active")
             }
             Route::GetAuditLogs {
                 action_type,
@@ -849,6 +891,69 @@ impl Display for RouteDisplay<'_> {
 
                 f.write_str("/pins")
             }
+            Route::GetJoinedPrivateArchivedThreads {
+                before,
+                channel_id,
+                limit,
+            } => {
+                f.write_str("channels/")?;
+                Display::fmt(channel_id, f)?;
+                f.write_str("/users/@me/threads/archived/private?")?;
+
+                if let Some(before) = before {
+                    f.write_str("before=")?;
+                    Display::fmt(before, f)?;
+                }
+
+                if let Some(limit) = limit {
+                    f.write_str("&limit=")?;
+                    Display::fmt(limit, f)?;
+                }
+
+                Ok(())
+            }
+            Route::GetPrivateArchivedThreads {
+                before,
+                channel_id,
+                limit,
+            } => {
+                f.write_str("channels/")?;
+                Display::fmt(channel_id, f)?;
+                f.write_str("/threads/archived/private?")?;
+
+                if let Some(before) = before {
+                    f.write_str("before=")?;
+                    Display::fmt(before, f)?;
+                }
+
+                if let Some(limit) = limit {
+                    f.write_str("&limit=")?;
+                    Display::fmt(limit, f)?;
+                }
+
+                Ok(())
+            }
+            Route::GetPublicArchivedThreads {
+                before,
+                channel_id,
+                limit,
+            } => {
+                f.write_str("channels/")?;
+                Display::fmt(channel_id, f)?;
+                f.write_str("/threads/archived/public")?;
+
+                if let Some(before) = before {
+                    f.write_str("before=")?;
+                    Display::fmt(before, f)?;
+                }
+
+                if let Some(limit) = limit {
+                    f.write_str("&limit=")?;
+                    Display::fmt(limit, f)?;
+                }
+
+                Ok(())
+            }
             Route::GetReactionUsers {
                 after,
                 channel_id,
@@ -881,6 +986,12 @@ impl Display for RouteDisplay<'_> {
 
                 Display::fmt(sticker_id, f)
             }
+            Route::GetThreadMembers { channel_id } => {
+                f.write_str("channels/")?;
+                Display::fmt(channel_id, f)?;
+
+                f.write_str("/thread-members")
+            }
             Route::GetUserConnections => f.write_str("users/@me/connections"),
             Route::GetUser { user_id } => {
                 f.write_str("users/")?;
@@ -898,6 +1009,12 @@ impl Display for RouteDisplay<'_> {
                 f.write_str(interaction_token)?;
 
                 f.write_str("/callback")
+            }
+            Route::JoinThread { channel_id } | Route::LeaveThread { channel_id } => {
+                f.write_str("channels/")?;
+                Display::fmt(channel_id, f)?;
+
+                f.write_str("/thread-members/@me")
             }
             Route::LeaveGuild { guild_id } => {
                 f.write_str("users/@me/guilds/")?;